+++ dart/geobase/example/geobase_example.dart
@@ -179,13 +179,25 @@ void _intro() {
   LineString.build([30, 10, 10, 30, 40, 40], type: Coords.xy);
   LineString.build([30, 10, 5.5, 10, 30, 5.5, 40, 40, 5.5], type: Coords.xyz);
 
+  // Position iterables can be used for building geomtries too.
+  LineString.from([
+    Projected(x: 30, y: 10),
+    Projected(x: 10, y: 30),
+    Projected(x: 40, y: 40),
+  ]);
+  LineString.from([
+    Geographic(lon: 30, lat: 10, elev: 5.5), // x = lon, y = lat, z = elev
+    Geographic(lon: 10, lat: 30, elev: 5.5),
+    Geographic(lon: 40, lat: 40, elev: 5.5),
+  ]);
+
   // Geometries can be parsed also from the text representation of coordinates.
   LineString.parseCoords('[30,10],[10,30],[40,40]'); // 2D
   LineString.parseCoords('[30,10,5.5],[10,30,5.5],[40,40,5.5]'); // 3D
 
   // -------
 
-  // GeoJSON and WKB formats are supported as input and output (WKT only output)
+  // GeoJSON, WKT and WKB formats are supported as input and output.
 
   // Parse a geometry from GeoJSON text.
   final geometry = LineString.parse(
@@ -544,7 +556,28 @@ void _geometryTypesOneliners() {
 
 void _point() {}
 
-void _lineString() {}
+void _lineString() {
+  // A line string with 3 points (2D coordinates with x and y).
+  LineString.from([
+    Projected(x: 30, y: 10),
+    Projected(x: 10, y: 30),
+    Projected(x: 40, y: 40),
+  ]);
+
+  // A line string with 3 points (3D coordinates with x, y and z).
+  LineString.from([
+    Geographic(lon: 30, lat: 10, elev: 5.5), // x = lon, y = lat, z = elev
+    Geographic(lon: 10, lat: 30, elev: 5.5),
+    Geographic(lon: 40, lat: 40, elev: 5.5),
+  ]);
+
+  // A line string with 3 points (3D and measured coordinates: x, y, z and m).
+  LineString.from([
+    Projected(x: 30, y: 10, z: 5.5, m: 3.1),
+    Projected(x: 10, y: 30, z: 5.5, m: 3.2),
+    Projected(x: 40, y: 40, z: 5.5, m: 3.3),
+  ]);
+}
 
 void _polygon() {}