+++ dart/geobase/lib/geobase.dart
@@ -96,7 +96,7 @@ export 'src/vector/formats/wkt/wkt_like_format.dart';
 
 // vector_data
 export 'src/vector_data/array/coordinates.dart';
-export 'src/vector_data/array/list_coordinate_extension.dart';
+export 'src/vector_data/array/coordinates_extensions.dart';
 export 'src/vector_data/coords/lonlat.dart';
 export 'src/vector_data/coords/xy.dart';
 export 'src/vector_data/model/bounded/bounded.dart';