+++ dart/geobase/lib/src/vector_data/model/geometry/multi_linestring.dart
@@ -10,6 +10,8 @@ import 'dart:typed_data';
 import '/src/codes/coords.dart';
 import '/src/codes/geom.dart';
 import '/src/constants/epsilon.dart';
+import '/src/coordinates/base/box.dart';
+import '/src/coordinates/base/position.dart';
 import '/src/coordinates/crs/coord_ref_sys.dart';
 import '/src/coordinates/projection/projection.dart';
 import '/src/utils/coord_arrays.dart';
@@ -22,6 +24,7 @@ import '/src/vector/formats/geojson/default_format.dart';
 import '/src/vector/formats/geojson/geojson_format.dart';
 import '/src/vector/formats/wkb/wkb_format.dart';
 import '/src/vector_data/array/coordinates.dart';
+import '/src/vector_data/array/coordinates_extensions.dart';
 
 import 'geometry.dart';
 import 'geometry_builder.dart';
@@ -38,13 +41,29 @@ class MultiLineString extends SimpleGeometry {
   ///
   /// An optional [bounds] can used set a minimum bounding box for a geometry.
   ///
-  /// Each line string or a chain of positions is represented by [PositionArray]
-  /// instances.
+  /// Each line string or a chain of positions is represented by a
+  /// [PositionArray] instance.
   const MultiLineString(List<PositionArray> lineStrings, {BoxCoords? bounds})
       : this._(lineStrings, bounds: bounds);
 
   const MultiLineString._(this._lineStrings, {super.bounds, Coords? type})
       : _type = type;
+
+  /// A multi line string from an iterable of [lineStrings] (each a chain as an
+  /// iterable of positions).
+  ///
+  /// An optional [bounds] can used set a minimum bounding box for a geometry.
+  ///
+  /// Each line string or a chain of positions is represented by an
+  /// `Iterable<Position>` instance.
+  factory MultiLineString.from(
+    Iterable<Iterable<Position>> lineStrings, {
+    Box? bounds,
+  }) =>
+      MultiLineString._(
+        lineStrings.map((chain) => chain.array()).toList(growable: false),
+        bounds: bounds?.coords(),
+      );
 
   /// Builds a multi line string from an array of [lineStrings] (each with a
   /// chain of positions).
@@ -54,8 +73,8 @@ class MultiLineString extends SimpleGeometry {
   ///
   /// An optional [bounds] can used set a minimum bounding box for a geometry.
   ///
-  /// Each line string or a chain of positions is represented by `
-  /// Iterable<double>` instances. They contain coordinate values as a flat
+  /// Each line string or a chain of positions is represented by a
+  /// `Iterable<double>` instance. They contain coordinate values as a flat
   /// structure. For example for `Coords.xyz` the first three coordinate values
   /// are x, y and z of the first position, the next three coordinate values are
   /// x, y and z of the second position, and so on.