+++ dart/geobase/lib/src/vector_data/model/geometry/point.dart
@@ -10,6 +10,7 @@ import 'dart:typed_data';
 import '/src/codes/coords.dart';
 import '/src/codes/geom.dart';
 import '/src/constants/epsilon.dart';
+import '/src/coordinates/base/position.dart';
 import '/src/coordinates/crs/coord_ref_sys.dart';
 import '/src/coordinates/projection/projection.dart';
 import '/src/utils/coord_arrays.dart';
@@ -21,6 +22,7 @@ import '/src/vector/formats/geojson/default_format.dart';
 import '/src/vector/formats/geojson/geojson_format.dart';
 import '/src/vector/formats/wkb/wkb_format.dart';
 import '/src/vector_data/array/coordinates.dart';
+import '/src/vector_data/array/coordinates_extensions.dart';
 
 import 'geometry.dart';
 import 'geometry_builder.dart';
@@ -31,6 +33,9 @@ class Point implements SimpleGeometry {
 
   /// A point geometry with [position].
   const Point(PositionCoords position) : _position = position;
+
+  /// A point geometry from [position].
+  factory Point.from(Position position) => Point(position.coords());
 
   /// Builds a point geometry from a [position].
   ///