+++ dart/geobase/lib/src/vector_data/model/geometry/polygon.dart
@@ -10,6 +10,8 @@ import 'dart:typed_data';
 import '/src/codes/coords.dart';
 import '/src/codes/geom.dart';
 import '/src/constants/epsilon.dart';
+import '/src/coordinates/base/box.dart';
+import '/src/coordinates/base/position.dart';
 import '/src/coordinates/crs/coord_ref_sys.dart';
 import '/src/coordinates/projection/projection.dart';
 import '/src/utils/coord_arrays.dart';
@@ -22,6 +24,7 @@ import '/src/vector/formats/geojson/default_format.dart';
 import '/src/vector/formats/geojson/geojson_format.dart';
 import '/src/vector/formats/wkb/wkb_format.dart';
 import '/src/vector_data/array/coordinates.dart';
+import '/src/vector_data/array/coordinates_extensions.dart';
 
 import 'geometry.dart';
 import 'geometry_builder.dart';
@@ -37,7 +40,7 @@ class Polygon extends SimpleGeometry {
   ///
   /// An optional [bounds] can used set a minimum bounding box for a geometry.
   ///
-  /// Each ring in the polygon is represented by `PositionArray` instances.
+  /// Each ring in the polygon is represented by a `PositionArray` instance.
   ///
   /// An empty polygon has no rings.
   ///
@@ -60,6 +63,30 @@ class Polygon extends SimpleGeometry {
         );
         */
 
+  /// A polygon geometry with one exterior and 0 to N interior [rings].
+  ///
+  /// An optional [bounds] can used set a minimum bounding box for a geometry.
+  ///
+  /// Each ring in the polygon is represented by an `Iterable<Position>`
+  /// instance.
+  ///
+  /// An empty polygon has no rings.
+  ///
+  /// For "normal" polygons the [rings] list must be non-empty. The first
+  /// element is the exterior ring, and any other rings are interior rings (or
+  /// holes). All rings must be closed linear rings. As specified by GeoJSON,
+  /// they should "follow the right-hand rule with respect to the area it
+  /// bounds, i.e., exterior rings are counterclockwise, and holes are
+  /// clockwise".
+  factory Polygon.from(
+    Iterable<Iterable<Position>> rings, {
+    Box? bounds,
+  }) =>
+      Polygon._(
+        rings.map((ring) => ring.array()).toList(growable: false),
+        bounds: bounds?.coords(),
+      );
+
   /// Builds a polygon geometry from one exterior and 0 to N interior [rings].
   ///
   /// Use [type] to specify the type of coordinates, by default `Coords.xy` is
@@ -67,8 +94,8 @@ class Polygon extends SimpleGeometry {
   ///
   /// An optional [bounds] can used set a minimum bounding box for a geometry.
   ///
-  /// Each ring in the polygon is represented by `Iterable<double>` arrays. Such
-  /// arrays contain coordinate values as a flat structure. For example for
+  /// Each ring in the polygon is represented by an `Iterable<double>` array.
+  /// Such arrays contain coordinate values as a flat structure. For example for
   /// `Coords.xyz` the first three coordinate values are x, y and z of the first
   /// position, the next three coordinate values are x, y and z of the second
   /// position, and so on.