+++ dart/geobase/test/vector_data/position_coords_test.dart
@@ -94,6 +94,31 @@ void main() {
       expect(xyzPosFromXym.optZ, isNotNull);
     });
 
+    test('Coords from extension', () {
+      const xyData = [15.0, 30.1];
+      const xyzData = [15.0, 30.1, 45.2];
+      const xymData = [15.0, 30.1, 60.3];
+      const xyzmData = [15.0, 30.1, 45.2, 60.3];
+
+      expect(PositionCoords.view(xyData), Projected.build(xyData).coords());
+      expect(
+        PositionCoords.view(xyzData, type: Coords.xyz),
+        Projected.build(xyzData).coords(),
+      );
+      expect(
+        PositionCoords.view(xyzData, type: Coords.xyz),
+        Projected.build(xyzData, type: Coords.xyz).coords(),
+      );
+      expect(
+        PositionCoords.view(xymData, type: Coords.xym),
+        Projected.build(xymData, type: Coords.xym).coords(),
+      );
+      expect(
+        PositionCoords.view(xyzmData, type: Coords.xyzm),
+        Projected.build(xyzmData, type: Coords.xyzm).coords(),
+      );
+    });
+
     test('Coordinate access and factories', () {
       final p1 = XY.create(x: 1.0, y: 2.0);
       final p2 = XYZ.create(x: 1.0, y: 2.0, z: 3.0);