+++ dart/geobase/CHANGELOG.md
@@ -15,6 +15,7 @@ NOTE: Version 0.5.0 currently [under development](https://github.com/navibyte/ge
 * [GeoJSON output for non-default coordinate reference systems #179](https://github.com/navibyte/geospatial/issues/179)
 * [WKT decoder on geobase #159](https://github.com/navibyte/geospatial/issues/159)
 * [Consistent equals, equals2D, equals3D (and equalsBy) for coordinates, geometries and features #146](https://github.com/navibyte/geospatial/issues/146)
+* [Geometry objects instantiated from instances or iterables of Position objects #171](https://github.com/navibyte/geospatial/issues/171)
 
 🛠 Refactoring:
 * [Enforce "geometry" under "Feature" when encoding GeoJSON text output #91](https://github.com/navibyte/geospatial/issues/91)